moves constructor
